+++ src/main/java/me/nikl/mpgamebundle/rockpaperscissors/RpsRules.java
@@ -9,6 +9,8 @@ import me.nikl.gamebox.game.rules.GameRuleRewards;
 public class RpsRules extends GameRuleRewards {
     private int numberOfRounds;
     private boolean looseOnTimeOver;
+    private int chooseTime = 10000;
+    private int waitTime = 5000;
 
     public RpsRules(String key, boolean saveStats, SaveType saveType, double cost, double moneyToWin, int tokenToWin, boolean looseOnTimeOver, int numberOfRounds) {
         super(key, saveStats, saveType, cost, moneyToWin, tokenToWin);
@@ -22,5 +24,21 @@ public class RpsRules extends GameRuleRewards {
 
     public boolean isLooseOnTimeOver() {
         return looseOnTimeOver;
+    }
+
+    public int getChooseTime() {
+        return chooseTime;
+    }
+
+    public void setChooseTime(int chooseTime) {
+        this.chooseTime = chooseTime;
+    }
+
+    public int getWaitTime() {
+        return waitTime;
+    }
+
+    public void setWaitTime(int waitTime) {
+        this.waitTime = waitTime;
     }
 }